requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pydantic
python-dotenv
sqlalchemy